OptionalExpandIntoPipe use node degree and caches found relationships
Refactored code so that `OptionalExpandIntoPipe` uses the same optimizations
as `ExpandIntoPipe`, i.e. look at degree and denseness to decide which node
to start from and cache interconnecting relationships.
